Minor Tweaks
* App folder structure update
* Added auth function to generate tokens
* README update
